feat: KPI-127 refactor CompanieTable
